Remove NS prefixes and _stubs suffixes
