+++ renderer-macos/lib/bindings/BriskCocoa.re
@@ -1,3 +1,3 @@
+type application;
 type window;
 type view;
-type application;
