+++ renderer-macos/lib/bindings/BriskImage.re
@@ -1,4 +1,4 @@
-type t = NSView.t;
+type t = BriskCocoa.view;
 
 type system = [
   | `ActionTemplate
