+++ renderer-macos/lib/components/View.re
@@ -10,7 +10,7 @@ let make = (~style: style=[], children) => {
   ...component,
   render: _ => {
     make: () => {
-      let view = NSView.make();
+      let view = BriskView.make();
       {view, layoutNode: makeLayoutNode(~style, view)};
     },
     shouldReconfigureInstance: (~oldState as _, ~newState as _) => true,
@@ -19,13 +19,13 @@ let make = (~style: style=[], children) => {
       |> List.iter(attr =>
            switch (attr) {
            | `Background(({r, g, b, a}: Color.t)) =>
-             NSView.setBackgroundColor(view, r, g, b, a)
+             BriskView.setBackgroundColor(view, r, g, b, a)
            | `Border(({width, color}: Border.t)) =>
              if (!isUndefined(width)) {
-               NSView.setBorderWidth(view, width);
+               BriskView.setBorderWidth(view, width);
              };
              let {r, g, b, a}: Color.t = color;
-             NSView.setBorderColor(view, r, g, b, a);
+             BriskView.setBorderColor(view, r, g, b, a);
            | #Layout.style => ()
            }
          );
