+++ renderer-macos/lib/stubs/BriskMenu.c
@@ -1,4 +1,4 @@
-#import "Cocoa_stubs.h"
+#import "BriskCocoa.h"
 
 enum {
   // Apple
