+++ renderer-macos/lib/stubs/BriskTextView.c
@@ -1,4 +1,4 @@
-#import "Cocoa_stubs.h"
+#import "BriskCocoa.h"
 
 @interface BriskTextView : NSTextView
