+++ renderer-macos/lib/stubs/BriskView.c
@@ -1,7 +1,4 @@
-#import "Cocoa_stubs.h"
-
-#define Val_View(v) ((value)(v))
-#define View_val(v) ((__bridge NSView *)(value)(v))
+#import "BriskCocoa.h"
 
 NSView *ml_NSView_make() {
   NSView *view = [NSView new];
